lucia-auth fixes for native
