+++ packages/app/features/sign-in/screen.native.tsx
@@ -1,7 +1,7 @@
-import { AuthProviderName } from '@t4/api/src/auth/shared'
+import type { AuthProviderName } from '@t4/api/src/auth/providers'
 import { YStack, useToastController } from '@t4/ui'
 import { SignUpSignInComponent } from 'app/features/sign-in/SignUpSignIn'
-import { useSignIn } from 'app/utils/auth'
+import { useSessionRedirect, useSignIn } from 'app/utils/auth'
 import { initiateAppleSignIn } from 'app/utils/auth/appleAuth'
 import { storeSessionToken } from 'app/utils/auth/credentials'
 import { isExpoGo } from 'app/utils/flags'
@@ -12,15 +12,17 @@ import * as WebBrowser from 'expo-web-browser'
 import { Platform } from 'react-native'
 
 export const SignInScreen = (): React.ReactNode => {
-  const { signIn } = useSignIn()
   const toast = useToastController()
   const utils = trpc.useUtils()
+  const { signIn } = useSignIn()
+
+  // Redirects back to the home page if signed in
+  useSessionRedirect({ true: '/' })
 
   const postLogin = () => {
     utils.user.invalidate()
     utils.auth.invalidate()
   }
-
 
   const signInWithApple = async () => {
     try {
